Upgrade to Airlift 0.78-SNAPSHOT api changes

+++ server/src/test/java/io/airlift/discovery/TestingTimeProvider.java
@@ -28,7 +28,7 @@ class TestingTimeProvider
 
     public void add(Duration interval)
     {
-        currentTime.addAndGet((long) interval.toMillis());
+        currentTime.addAndGet(interval.toMillis());
     }
 
     public void set(DateTime currentTime)
